Day033 Completed
Weekend of travel and vacation back on the grind however.

Today was a simple application to email me when the ISS is above based on my latitude and longitude.

Also learnt how to use APIs using the requests library and how to receive the data back from them in terms of JSON data